+++ Sources/RVS_Generic_Swift_Toolbox/Extensions/RVS_Int_Extensions.swift
@@ -88,26 +88,6 @@ public extension UInt64 {
         // By masking out anything not in the run length, we return a value.
         return shifted & UInt64(mask)
     }
-    
-    /* ################################################################## */
-    /**
-     This returns the Int as a Roman numeral.
-     NOTE: Values less than 1 are returned as "".
-     
-     From here: https://stackoverflow.com/a/36068105/879365
-     */
-    var romanNumeral: String {
-        var integerValue = self
-        var numeralString = ""
-        let mappingList: [(UInt64, String)] = [(1000, "M"), (900, "CM"), (500, "D"), (400, "CD"), (100, "C"), (90, "XC"), (50, "L"), (40, "XL"), (10, "X"), (9, "IX"), (5, "V"), (4, "IV"), (1, "I")]
-        for i in mappingList {
-            while (integerValue >= i.0) {
-                integerValue -= i.0
-                numeralString += i.1
-            }
-        }
-        return numeralString
-    }
 }
 
 /* ###################################################################################################################################### */
@@ -159,13 +139,6 @@ public extension UInt {
      - returns: An Int, with the masked value.
      */
     func maskedValue(firstPlace inFirstPlace: UInt, runLength inRunLength: UInt) -> UInt { UInt(UInt64(self).maskedValue(firstPlace: inFirstPlace, runLength: inRunLength)) }
-    
-    /* ################################################################## */
-    /**
-     This returns the Int as a Roman numeral.
-     NOTE: Values less than 1 are returned as "".
-     */
-    var romanNumeral: String { UInt64(self).romanNumeral }
 }
 
 /* ###################################################################################################################################### */
@@ -228,8 +201,21 @@ public extension Int {
     /* ################################################################## */
     /**
      This returns the Int as a Roman numeral.
-     NOTE: Values less than 1 are returned as "".
-     This uses an absolute value, so all returns are positive.
+     NOTE: The maximum value is 4999. The minimum value is 1. Out of range values are returned as "".
+
+     Inspired by the SO answer here: https://stackoverflow.com/a/36068105/879365
      */
-    var romanNumeral: String { UInt64(abs(self)).romanNumeral }
+    var romanNumeral: String {
+        guard (1..<5000).contains(self) else { return "" }
+        var integerValue = UInt16(self)
+        var numeralString = ""
+        let mappingList: [(UInt16, String)] = [(1000, "M"), (900, "CM"), (500, "D"), (400, "CD"), (100, "C"), (90, "XC"), (50, "L"), (40, "XL"), (10, "X"), (9, "IX"), (5, "V"), (4, "IV"), (1, "I")]
+        mappingList.forEach {
+            while (integerValue >= $0.0) {
+                integerValue -= $0.0
+                numeralString += $0.1
+            }
+        }
+        return numeralString
+    }
 }